pyproject.toml
[tool.poetry]
name = "neobot"
version = "0.1.0"
description = "A modern extensible discord bot written in python with discord.py"
authors = ["WizzyGeek"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "neobot", from = "."}
]


[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.0"
asyncpg = "~0"
DiscTools = {git = "https://github.com/WizzyBots/DiscTools"}
aiohttp-jinja2 = {version = "^1.4.2", optional = true}
aiohttp = {extras = ["speedups"], version = "^3.7.4"}
jishaku = "*"
typer = {extras = ["all"], version = "^0.3"}
polyleven = "^0.7"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
mypy = "^0.942"
asyncpg-stubs = {git = "https://github.com/bryanforbes/asyncpg-stubs.git"}
yappi = {version = "^1.3.2"}

[tool.poetry.extras]
web = ["aiohttp-jinja2"]

[tool.poetry.scripts]
neobot = "neobot.__main__:App"
neotools = "neobot.devtools:app"

[tool.pyright]
include = ["neobot"]
exclude = ["neobot/web/resources", ".venv", ".venv_39", ".venv_310"]
pythonVersion = "^3.8"
# venv = ".venv"
# venvPath = "./.venv"
# Shut up pylance I know what i am doing
reportGeneralTypeIssues = false
reportOptionalMemberAccess = false
reportInvalidStringEscapeSequence = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
